Moving reflction util to helper dir
